+++ finance_news/__init__.py
@@ -1,0 +1,26 @@
+"""Finance news MCP package."""
+
+from .network import SESSION, _http_get
+from .data_sources import (
+    COMMODITY_MAP,
+    FX_ALIAS,
+    INDEX_MAP,
+    EQUITY_MAP,
+    _fetch_yahoo_chart,
+    _yahoo_options_chain,
+    _news_all,
+)
+from .tools import app
+
+__all__ = [
+    "app",
+    "SESSION",
+    "_http_get",
+    "COMMODITY_MAP",
+    "FX_ALIAS",
+    "INDEX_MAP",
+    "EQUITY_MAP",
+    "_fetch_yahoo_chart",
+    "_yahoo_options_chain",
+    "_news_all",
+]
